Create Ludo
There is a rule in ludo that a token can only be unlocked when either a 1 or 6 shown in the die. Given the die number N, Your task is to check whether the token can be unlocked or not.
Input
User Task:
Since this will be a functional problem, you don't have to take input. You just have to complete the function ludo() that takes integers N as argument.

Constraints:-
1 <= N <= 6
Output
Return 1 if the token can be unlocked else return 0.