Create leap year
Given a year, find if it is a leap year. Leap year is the year that is multiple of 4. But, multiples of 100 which are not multiples of 400 are not leap years.
Input
User Task:
Since this will be a functional problem, you don't have to take input. You just have to complete the function LeapYear() that takes integer n as a parameter.

Constraint:
1 <= n <= 5000
Output
Return 1 if the year is a leap year else return 0.