Create Rook
Given an 8*8 empty chessboard in which a rook is placed at a position (X, Y). Your task is to find the minimum steps Rook will take to go to the position (1, 1).
Input
User Task:
Since this will be a functional problem, you don't have to take input. You just have to complete the function Rook() that takes integers X and Y as arguments.

Constraints:-
1 <= X <= 8
1 <= Y <= 8
Output
Return the number of steps rook will take to go to the position (X, Y).
Example